PR comments: Move utils to appropriate class, fix word selection algorithm, fix the parallelism
